% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getClusterResults}
\alias{getClusterResults}
\title{getClusterResults}
\usage{
getClusterResults(connectionDetails, cdmDatabaseSchema)
}
\arguments{
\item{connectionDetails}{The connect details to connect to the database}

\item{cdmDatabaseSchema}{Thew schema to perform the clustering on}
}
\value{
A list containing the individual cluster results, the similarity matrix and the condition concept ids included
}
\description{
This returns a list containing the individual cluster results, the similarity matrix and the condition concept ids included
}
\details{
This performs the ensemble clustering on your data


Users input a cdmDatabaseSchema and connectionDetails then the data are extracted from the CDM and clustering performed
}
