% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getSimilarConcepts}
\alias{getSimilarConcepts}
\title{GetSimilarConditions}
\usage{
getSimilarConcepts(conditionConceptId, n = 50)
}
\arguments{
\item{conditionConceptId}{The concept id for the condition you want similar conditions to}

\item{n}{The number of similar conditions return (default is 50 - needs to be less than 50)}
}
\value{
A dataframe with the similar condition names, concept ids and similarity measure
}
\description{
This returns a ranked list of similar conditions for a user input condition concept_id
}
\details{
This returns a ranked list of similar conditions for a user input condition concept_id


Users input a condition concept id and get a ranked list lf similar conditions returned
}
