% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{includedConditions}
\alias{includedConditions}
\title{Included condition concepts}
\format{A data frame with 30652 rows and 2 variables:
\describe{
  \item{CONDITION_CONCEPT_ID}{The concept id of the condition}
  \item{CONCEPT_NAME}{The corresponding name}
}}
\usage{
includedConditions
}
\description{
A dataset containing the condition concept ids and names for those included
in this package
}
\keyword{datasets}
