% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{similarityTop50}
\alias{similarityTop50}
\title{The top 50 similar condition concepts for each condition concept id}
\format{A data frame with 30652 rows and 2 variables:
\describe{
  \item{CONCEPT_ID_OI}{The concept id of the condition of interest}
  \item{CONDITION_CONCEPT_ID}{The concept id of the similar condition concept}
  \item{CONCEPT_NAME}{The corresponding name of the similar condition concept}
  \item{CONDITION_CONCEPT_ID}{The number of clusters that the concept of interest has in common with the similar concept}
}}
\usage{
similarityTop50
}
\description{
A dataset containing the condition concept ids and the top 50 similar condition
concept ids
}
\keyword{datasets}
